neurons/miners/openai/requirements.txt
openai==1.9.0
langchain==0.1.0
python-dotenv
